Fixed reconnect issue

+++ src/main/java/com/lks/core/FTPUploader.java
@@ -5,9 +5,11 @@ import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.net.PrintCommandListener;
 import org.apache.commons.net.ftp.FTP;
 import org.apache.commons.net.ftp.FTPClient;
+import org.apache.commons.net.ftp.FTPConnectionClosedException;
 import org.apache.commons.net.ftp.FTPReply;
 
 import java.io.*;
+import java.net.SocketException;
 
 /**
  * Created by lokkur on 8/18/2015.
@@ -15,22 +17,48 @@ import java.io.*;
 public class FTPUploader {
     FTPClient ftp = null;
 
+    private String host;
+    private String user;
+    private String pwd;
+
     public FTPUploader(String host, String user, String pwd) throws Exception{
         ftp = new FTPClient();
-        ftp.addProtocolCommandListener(new PrintCommandListener(new PrintWriter(System.out)));
-        int reply;
-        ftp.connect(host);
-        reply = ftp.getReplyCode();
-        if (!FTPReply.isPositiveCompletion(reply)) {
-            ftp.disconnect();
-            throw new FALException(ExceptionCode.SYSTEM_ERROR,"Exception in connecting to FTP Server");
-        }
-        ftp.login(user, pwd);
-        ftp.setFileType(FTP.BINARY_FILE_TYPE);
-        ftp.enterLocalPassiveMode();
+        this.host = host;
+        this.user = user;
+        this.pwd = pwd;
+        connect();
     }
+
+    public String getHost() {
+        return host;
+    }
+
+    public void setHost(String host) {
+        this.host = host;
+    }
+
+    public String getUser() {
+        return user;
+    }
+
+    public void setUser(String user) {
+        this.user = user;
+    }
+
+    public String getPwd() {
+        return pwd;
+    }
+
+    public void setPwd(String pwd) {
+        this.pwd = pwd;
+    }
+
     public synchronized String uploadFile(String localFileFullName, String fileName, String hostDir){
         if(ftp.isConnected()){
+            if(!checkIfConnectionExists()){
+                throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to connect to FTP server");
+
+            }
             try(InputStream input = new FileInputStream(new File(localFileFullName))){
                 String fileLocation = hostDir + "/" + fileName;
                 boolean b = this.ftp.storeFile(fileLocation, input);
@@ -67,6 +95,10 @@ public class FTPUploader {
     public synchronized boolean checkDirectoryExists(String dirPath) {
         int returnCode;
         try {
+            if(!checkIfConnectionExists()){
+                throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to connect to FTP server");
+
+            }
             ftp.changeWorkingDirectory(dirPath);
 
         returnCode = ftp.getReplyCode();
@@ -83,6 +115,10 @@ public class FTPUploader {
 
     public synchronized boolean createDirectory(String dirPath){
         try {
+            if(!checkIfConnectionExists()){
+                throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to connect to FTP server");
+
+            }
             return ftp.makeDirectory(dirPath);
         } catch (IOException e) {
             throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to create new directory "+ dirPath);
@@ -92,6 +128,10 @@ public class FTPUploader {
 
     public synchronized boolean deleteFile(String fileLocation) {
         try {
+            if(!checkIfConnectionExists()){
+                throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to connect to FTP server");
+
+            }
             return ftp.deleteFile(fileLocation);
         } catch (IOException e) {
             throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to delete the file: "+ fileLocation);
@@ -102,6 +142,10 @@ public class FTPUploader {
         File tmpFile = null;
         String[] split = fileLocation.split("/");
         String fileName = split[split.length - 1];
+        if(!checkIfConnectionExists()){
+            throw new FALException(ExceptionCode.SYSTEM_ERROR, "Unable to connect to FTP server");
+
+        }
         try{
             tmpFile = File.createTempFile(FilenameUtils.getBaseName(fileName), "." + FilenameUtils.getExtension(fileName));
             OutputStream outputStream1 = new BufferedOutputStream(new FileOutputStream(tmpFile));
@@ -118,5 +162,108 @@ public class FTPUploader {
             throw new FALException(ExceptionCode.SYSTEM_ERROR, "Encountered exception when tryin to download file: "+ fileLocation);
         }
 
+    }
+
+
+
+    private synchronized boolean connect(){
+        try{
+            ftp.addProtocolCommandListener(new PrintCommandListener(new PrintWriter(System.out)));
+            int reply;
+            ftp.connect(host);
+            reply = ftp.getReplyCode();
+            if (!FTPReply.isPositiveCompletion(reply)) {
+                ftp.disconnect();
+                throw new FALException(ExceptionCode.SYSTEM_ERROR,"Exception in connecting to FTP Server");
+            }
+            ftp.login(user, pwd);
+            ftp.setFileType(FTP.BINARY_FILE_TYPE);
+            ftp.setDefaultTimeout(6000);
+            ftp.setConnectTimeout(3000);
+            ftp.setSoTimeout(3000);
+            ftp.enterLocalPassiveMode();
+            return true;
+        } catch (SocketException e) {
+            throw new FALException("Unable to connect to FTP server");
+        } catch (IOException e) {
+            throw new FALException("Unable to connect to FTP server");
+        }
+
+    }
+
+    private boolean checkIfConnectionExists(){
+
+        try
+        {
+            // Sends a NOOP command to the FTP server.
+            boolean answer = ftp.sendNoOp();
+            if(answer)
+                return true;
+            else
+            {
+                System.out.println("Server connection failed!");
+                boolean success = connect();
+                if(success)
+                {
+                    System.out.println("Reconnect attampt have succeeded!");
+                    return true;
+                }
+                else
+                {
+                    System.out.println("Reconnect attampt failed!");
+                    disconnect();
+                    return false;
+                }
+            }
+        }
+        catch (FTPConnectionClosedException e)
+        {
+            System.out.println("Server connection is closed!");
+            boolean recon = connect();
+            if(recon)
+            {
+                System.out.println("Reconnect attampt have succeeded!");
+                return true;
+            }
+            else
+            {
+                System.out.println("Reconnect attampt have failed!");
+                disconnect();
+                return false;
+            }
+
+        }
+        catch (IOException e)
+        {
+            System.out.println("Server connection failed!");
+            boolean recon = connect();
+            if(recon)
+            {
+                System.out.println("Reconnect attampt have succeeded!");
+                return true;
+            }
+            else
+            {
+                System.out.println("Reconnect attampt have failed!");
+                disconnect();
+                return false;
+            }
+        }
+        catch (NullPointerException e)
+        {
+            System.out.println("Server connection is closed!");
+            boolean recon = connect();
+            if(recon)
+            {
+                System.out.println("Reconnect attampt have succeeded!");
+                return true;
+            }
+            else
+            {
+                System.out.println("Reconnect attampt have failed!");
+                disconnect();
+                return false;
+            }
+        }
     }
 }
